DAY 76Project 15, part 3
Challenge 2: Hints and labelsThe second challenge is to fix the list rows in iExpense so they read out the name and value in one single VoiceOver label, and their type in a hint. Right now these rows are read quite badly, with each of those things being read separately inside the row – as you’re swiping through several rows it becomes hard to remember which value belongs where.